unintended programs work around
it seams like undesired programs for ID4, 10, and 11 were introduced. I'm not sure if this was due to the IP days fixes or something else. I've put in a workaround to resolve the issue, however the fix is not good enough IMO. A more general fix limiting programs should be considered and likely applied in the saving layer. Sub indicator queries could retain all programs without harm then simplifying filters and making additions and revisions easier.